Prism runs properly.
The specification patterns still need to be rewritten to match prism
syntax, but the model checker does run.
NOTE: in order to get the prism batch file to run, you have to manually
change the PRISM_DIR= field in the .bat file to the absolute path to the
compiled prism source.  In the future it might be possible to run prism
by jar packages rather than by command line.